+++ lib/src/posthog_constants.dart
@@ -1,0 +1,14 @@
+/// Event names used throughout the PostHog Flutter SDK
+class PostHogEventName {
+  PostHogEventName._();
+
+  static const screen = '\$screen';
+  static const exception = '\$exception';
+}
+
+/// Property keys used throughout the PostHog Flutter SDK
+class PostHogPropertyName {
+  PostHogPropertyName._();
+
+  static const screenName = '\$screen_name';
+}